move Object to runtime::object::Object
